Test fixing types tva

+++ internal/models/menu_models.go
@@ -39,9 +39,9 @@ type ProductEntry struct {
 	Price                        int64                `json:"price"`
 	PriceTakeAway                int64                `json:"price_take_away"`
 	PriceDelivery                int64                `json:"price_delivery"`
-	TVAIn                        int64                `json:"tva_rate_in,omitempty"`
-	TVADelivery                  int64                `json:"tva_rate_delivery,omitempty"`
-	TVATakeAway                  int64                `json:"tva_rate_take_away,omitempty"`
+	TVAIn                        float64              `json:"tva_rate_in,omitempty"`
+	TVADelivery                  float64              `json:"tva_rate_delivery,omitempty"`
+	TVATakeAway                  float64              `json:"tva_rate_take_away,omitempty"`
 	AvailableIn                  bool                 `json:"available_in,omitempty"`
 	AvailableTakeAway            bool                 `json:"available_take_away,omitempty"`
 	AvailableDelivery            bool                 `json:"available_delivery,omitempty"`
